refactor: make a poetry project
