+++ Makefile
@@ -1,23 +1,17 @@
-VENV?=.venv
-PYTHON=$(VENV)/bin/python3
-PYTEST=$(VENV)/bin/pytest
-PIP=$(PYTHON) -m pip
-FLASK=$(PYTHON) -m flask
-SHELL := /bin/bash
-
 help:
 	@echo "make [ venv | run | test | deploy ]"
 
-venv:
-	@echo "Starting virtual environment..."
-	python3 -m venv .venv
-	@echo "Installing requirements..."
-	$(PIP) install --upgrade pip
-	$(PIP) install -r requirements.txt
-	$(PIP) install -e .
+
+setup-poetry:
+	pip install poetry
+
+venv: setup-poetry
+	poetry install
+	pip install --upgrade pip
+	pip install -e src/
 
 run:
-	$(FLASK) --app src.main run
+	poetry run flask --app src.main.main run
 
 test:
-	$(PYTEST) -vv
+	poetry run pytest
