+++ pyproject.toml
@@ -1,0 +1,25 @@
+[tool.poetry]
+name = "facilitai"
+version = "1.0.0"
+description = ""
+authors = ["UFCG"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+flask = "2.2.3"
+flask-cors = "3.0.10"
+authlib = "1.2.0"
+sqlalchemy = "2.0.15"
+flask-sqlalchemy = "3.0.3"
+psycopg2 = "^2.9.6"
+flask-jwt-extended = "4.4.4"
+facilitai-package = { path = "./src", develop = true }
+
+[tool.poetry.group.dev.dependencies]
+pytest = "7.1.3"
+testcontainers = "3.7.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
