--- setup.py
@@ -1,8 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(
-    name='facilitai',
-    version='1.0.1',
-    packages=find_packages(),
-    package_dir={'': 'src'}
-)
