+++ src/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='facilitai-package',
+    version='1.0.1',
+    packages=find_packages()
+)
